NLP/Nepal-Constitution-Chatbot/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.26
langchain-groq>=0.3.2
huggingface-hub>=0.19.4
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
tqdm>=4.66.1
streamlit>=1.24.0